requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7
python-dateutil==2.8.2
typing-extensions==4.8.0
lxml==4.9.3
